pyproject.toml
[tool.poetry]
name = "virtualassistentbackend"
version = "0.1.0"
description = ""
authors = ["chacachien <nguyentienphatbl123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.110.2"
uvicorn = "0.29.0"
alembic = "1.13.1"
sqlmodel = "0.0.16"
sqlalchemy = "2.0.29"
databases = "0.9.0"
mysql-connector-python = "8.3.0"
python-dotenv = "1.0.1"
pydantic = "2.7.0"
pytz = "2024.1"
transformers = "4.39.1"
langchain = "0.1.13"
langchain-community = "0.0.34"
pinecone-client = "3.2.2"
sentence-transformers = "2.6.1"
starlette = "^0.37.2"
fastapi-sqlalchemy = "^0.2.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
langchain-google-genai = "^1.0.2"
psycopg2-binary = "^2.9.9"
langchain-openai = "^0.1.3"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
